backend/requirements.txt
# Backend FastAPI Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Data Processing (reusando lógica Mougli/Mapito)
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
xlsxwriter==3.1.9
pyarrow==14.0.2

# Mapas (Mapito)
folium==0.15.1
branca==0.7.1
jinja2==3.1.3

# Utils
python-dotenv==1.0.0
